# api/openapi.yaml
openapi: 3.1.0
info:
  title: Runix API
  version: 0.1.0
  description: |
    Control-plane API of the Runix infrastructure management platform.

    Authentication: `Authorization: Bearer <token>` where the token is either
    a JWT access token (from /auth/login or /auth/refresh) or a personal
    access token (`rnx_pat_...`). WebSocket endpoints accept the JWT via the
    `access_token` query parameter because browsers cannot set headers on
    WebSocket upgrades.

    Every error response has the shape
    `{"error": {"code": "...", "message": "..."}}`.
servers:
  - url: /api/v1
...
paths:
...
  /servers/{id}/runtimes:
    get:
      tags: [runtimes]
      summary: List runtime instances on a server (server.view)
      parameters:
        - { $ref: "#/components/parameters/id" }
        - { name: type, in: query, schema: { type: string, enum: [docker, compose, systemd, daemon] } }
      responses:
        "200":
          description: Runtime descriptors with per-provider capabilities
          content:
            application/json:
              schema:
                type: object
                properties:
                  runtimes:
                    type: array
                    items: { $ref: "#/components/schemas/RuntimeInfo" }
        "503": { description: Agent offline }
  /servers/{id}/runtimes/{type}:
    post:
      tags: [runtimes]
      summary: Create a runtime instance (runtime.manage or <type>.manage)
      description: |
        `config` is the provider-specific document. docker: image, cmd, env,
        binds, ports, restartPolicy. daemon: cmd, workingDir, env, autoStart,
        restartPolicy (never|on-failure|always), maxRestarts,
        restartDelaySeconds, stopSignal, stopTimeoutSeconds.
      parameters:
        - { $ref: "#/components/parameters/id" }
        - { $ref: "#/components/parameters/rtype" }
      responses:
        "201": { description: Created, returns the new instance id }
        "422": { description: Provider does not support create }
  /servers/{id}/runtimes/{type}/{rid}:
    get:
      tags: [runtimes]
      summary: Get one runtime instance (server.view)
      parameters:
        - { $ref: "#/components/parameters/id" }
        - { $ref: "#/components/parameters/rtype" }
        - { $ref: "#/components/parameters/rid" }
      responses:
        "200":
          description: Runtime info
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RuntimeInfo" }
    put:
      tags: [runtimes]
      summary: Update a runtime's configuration in place (runtime.manage or <type>.manage)
      description: |
        Only providers advertising the `update` capability accept this
        (daemons today). `config` is the same provider-specific document as
        create; a running daemon is restarted to apply. Returns 422 when the
        runtime cannot be reconfigured in place.
      parameters:
        - { $ref: "#/components/parameters/id" }
        - { $ref: "#/components/parameters/rtype" }
        - { $ref: "#/components/parameters/rid" }
      responses:
        "204": { description: Updated }
        "422": { description: Not supported by this runtime }
    delete:
      tags: [runtimes]
      summary: Remove a runtime instance (runtime.manage or <type>.manage)
      parameters:
        - { $ref: "#/components/parameters/id" }
        - { $ref: "#/components/parameters/rtype" }
        - { $ref: "#/components/parameters/rid" }
        - { name: force, in: query, schema: { type: boolean } }
        - { name: purge, in: query, schema: { type: boolean } }
      responses:
        "204": { description: Removed }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
  parameters:
    id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    rtype:
      name: type
      in: path
      required: true
      schema: { type: string, enum: [docker, compose, systemd, daemon] }
    rid:
      name: rid
      in: path
      required: true
      schema: { type: string }
...
  schemas:
...
    RuntimeInfo:
      type: object
      properties:
        descriptor:
          type: object
          properties:
            id: { type: string }
            type: { type: string }
            name: { type: string }
            labels: { type: object, additionalProperties: { type: string } }
            createdAt: { type: string, format: date-time }
            status:
              type: object
              properties:
                state:
                  type: string
                  enum: [created, starting, running, degraded, paused, stopping, stopped, failed, unknown]
                health:
                  type: string
                  enum: [unknown, starting, healthy, unhealthy]
                message: { type: string }
                exitCode: { type: integer }
                startedAt: { type: string, format: date-time }
                finishedAt: { type: string, format: date-time }
                restartCount: { type: integer }
        capabilities:
          type: array
          items: { type: string }
security:
  - bearer: []